# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAQ API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/control/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from current device",
                "description": "Logs out the authenticated user from the current device.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logged out from current device",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out from current device."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/logout-all": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Logs out the authenticated user from all devices.",
                "operationId": "logoutAll",
                "responses": {
                    "200": {
                        "description": "Logged out from all devices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out from all devices."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": [],
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
        "/api/control/roles/list": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get list of roles",
                "operationId": "89d42ac89b32436b22af9143b8f94d4c",
                "responses": {
                    "200": {
                        "description": "Roles list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RolesListResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    },
                    {
                        "SanctumBearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RolesListResource": {
                "title": "Roles List Resource",
                "description": "Roles list resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "guard_name": {
                        "type": "string",
                        "example": "web"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiToken": {
                "type": "apiKey",
                "description": "ApiToken required for all end points",
                "name": "Token",
                "in": "header"
            },
            "SanctumBearerToken": {
                "type": "http",
                "description": "Sanctum Bearer Token required after user logs in (Authorization: Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}